Update toggleCellLife to handle invalid rowIndex or columnIndex
toggleCellLife should return the original generation if the rowIndex or columnIndex arguments are invalid. 

If rowIndex is less than zero or greater than the index of the last row in the given generation, it is not a valid argument.

If columnIndex is less than zero or greater than the index of the last column in the given generation, it is not a valid argument.